Convert indices to signed integers
Although the semantics (signed == can be invalid; unsigned == can't)
were useful, the distinction that WGPU makes between them create
unwieldy code. Convert all unsigned indices to signed ones.

+++ render/src/push_curve.wgsl
@@ -36,49 +36,49 @@ fn push_curve_main(@builtin(global_invocation_id) global_invocation_id: vec3<u32
 ///  - The midpoint for a curve that has 2+ points
 fn push_vertex(pos: vec2<f32>, curve_index: u32) {
     var curve = curves[curve_index];
-    let index = atomicAdd(&vertices.size, 2u);
+    let index = atomicAdd(&vertices.size, 2);
 
     var new_ray = vec2<f32>(0.0, 0.0);
     if (curve.tail_index >= 0) {
         // If we have 2+ points, calculate the new ray based on the previous point
-        new_ray = calculate_new_rays(u32(curve.head_index), pos);
+        new_ray = calculate_new_rays(curve.head_index, pos);
 
         // Write the indices out
-        push_quad_indices(u32(curve.head_index), index);
+        push_quad_indices(curve.head_index, index);
 
         // Push the linked list
-        vertices.data[curve.head_index].next = i32(index);
+        vertices.data[curve.head_index].next = index;
     }
     else {
         // If this is our first point, indicate that
-        curve.tail_index = i32(index);
+        curve.tail_index = index;
     }
 
     // In all cases, write out the new point
-    vertices.data[index     ] = Vertex(pos,  new_ray, 1.0, -1);
-    vertices.data[index + 1u] = Vertex(pos, -new_ray, 1.0, -1);
+    vertices.data[index    ] = Vertex(pos,  new_ray, 1.0, -1);
+    vertices.data[index + 1] = Vertex(pos, -new_ray, 1.0, -1);
 
-    curve.head_index = i32(index);
+    curve.head_index = index;
     curves[curve_index] = curve;
 }
 
 /// Place the 6 indices (one quad) into the index buffer.
-fn push_quad_indices(i: u32, j: u32) {
-    let index = atomicAdd(&indices.size, 6u);
+fn push_quad_indices(i: i32, j: i32) {
+    let index = atomicAdd(&indices.size, 6);
 
     // 0, 1, 2
-    indices.data[index     ] = i;
-    indices.data[index + 1u] = i + 1u;
-    indices.data[index + 2u] = j;
+    indices.data[index    ] = i;
+    indices.data[index + 1] = i + 1;
+    indices.data[index + 2] = j;
 
     // 1, 3, 2
-    indices.data[index + 3u] = i + 1u;
-    indices.data[index + 4u] = j + 1u;
-    indices.data[index + 5u] = j;
+    indices.data[index + 3] = i + 1;
+    indices.data[index + 4] = j + 1;
+    indices.data[index + 5] = j;
 }
 
 /// Compute the rays for the current- and mid-point of the segment list.
-fn calculate_new_rays(prev_index: u32, new_pos: vec2<f32>) -> vec2<f32> {
+fn calculate_new_rays(prev_index: i32, new_pos: vec2<f32>) -> vec2<f32> {
     let prev_vertex = vertices.data[prev_index];
 
     let ray_between = new_pos - prev_vertex.pos;
@@ -93,8 +93,8 @@ fn calculate_new_rays(prev_index: u32, new_pos: vec2<f32>) -> vec2<f32> {
 
     if (prev_vertex.ray.x == 0.0 && prev_vertex.ray.y == 0.0) {
         // If the previous ray is invalid, use the same for both
-        vertices.data[prev_index     ].ray =  new_ray;
-        vertices.data[prev_index + 1u].ray = -new_ray;
+        vertices.data[prev_index    ].ray =  new_ray;
+        vertices.data[prev_index + 1].ray = -new_ray;
         return new_ray;
     }
 
@@ -116,8 +116,8 @@ fn calculate_new_rays(prev_index: u32, new_pos: vec2<f32>) -> vec2<f32> {
 
     let mul = dot(v, v - u) / det;
     let mid_ray = (w * mul + u);
-    vertices.data[prev_index     ].ray =  mid_ray;
-    vertices.data[prev_index + 1u].ray = -mid_ray;
+    vertices.data[prev_index    ].ray =  mid_ray;
+    vertices.data[prev_index + 1].ray = -mid_ray;
 
     return new_ray;
 }
